(PC-13519)[PRO] refactor: move constatns to core/Offers
