+++ pro/src/components/pages/Offers/Offers/OfferItem/OfferItem.jsx
@@ -7,7 +7,7 @@ import Icon from 'components/layout/Icon'
 import Thumb from 'components/layout/Thumb'
 import { isOfferDisabled } from 'components/pages/Offers/domain/isOfferDisabled'
 import StatusLabel from 'components/pages/Offers/Offer/OfferStatus/StatusLabel'
-import { OFFER_STATUS_SOLD_OUT } from 'components/pages/Offers/Offers/_constants'
+import { OFFER_STATUS_SOLD_OUT } from 'core/Offers/constants'
 import { computeVenueDisplayName } from 'repository/venuesService'
 import { Tag } from 'ui-kit'
 import { pluralize } from 'utils/pluralize'
