+++ pro/src/components/pages/Offers/Offers/OffersTableHead/OffersTableHead.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import { useSelector } from 'react-redux'
 
-import { ADMINS_DISABLED_FILTERS_MESSAGE } from '../_constants'
+import { ADMINS_DISABLED_FILTERS_MESSAGE } from 'core/Offers/constants'
 
 import StatusFiltersButton from './StatusFiltersButton'
