+++ pro/src/components/pages/Offers/domain/getOffersCountToDisplay.js
@@ -1,4 +1,4 @@
-import { MAX_OFFERS_TO_DISPLAY } from 'components/pages/Offers/Offers/_constants'
+import { MAX_OFFERS_TO_DISPLAY } from 'core/Offers/constants'
 
 export const getOffersCountToDisplay = offersCount =>
   offersCount <= MAX_OFFERS_TO_DISPLAY ? offersCount.toString() : '500+'
