+++ pro/src/components/pages/Offers/domain/isOfferDisabled.js
@@ -1,7 +1,7 @@
 import {
   OFFER_STATUS_PENDING,
   OFFER_STATUS_REJECTED,
-} from 'components/pages/Offers/Offers/_constants'
+} from 'core/Offers/constants'
 
 export const isOfferDisabled = status => {
   return [OFFER_STATUS_REJECTED, OFFER_STATUS_PENDING].includes(status)
