+++ pro/src/repository/pcapi/pcapi.js
@@ -1,9 +1,6 @@
 import { DEFAULT_PRE_FILTERS } from 'components/pages/Bookings/PreFilters/_constants'
-import {
-  ALL_OFFERERS,
-  DEFAULT_SEARCH_FILTERS,
-} from 'components/pages/Offers/Offers/_constants'
 import { DEFAULT_INVOICES_FILTERS } from 'components/pages/Reimbursements/_constants'
+import { ALL_OFFERERS, DEFAULT_SEARCH_FILTERS } from 'core/Offers/constants'
 import { client } from 'repository/pcapi/pcapiClient'
 import {
   FORMAT_ISO_DATE_ONLY,
